minor debug, starting league-wide stats
fixed printing of Lowest Margin of Victory. was using the wrong global
variable in the print statement

also started work on league-wide stats by collected all individual game
scores and their owners into an array called String[][]
ListOfGameScores. the plan is to go through this data and sift out the
top 10 and bottom 10, and then add them to the printing process