implemented top 5 / bottom 5 score tracking
this one was tough. decided to do it without any fancy data structures.
highestAndLowest() was created and gets called from inside massageData,
after all the loops. it uses the stored scores from ListOfGameScores[][]
and ends up storing the new data in leagueWideStats[][], which is going
to be the receptacle for ALL league-wide stats