DataManager first pass
ripped out nearly all the bad static variables and replaced them with a
new DataManager class. this is supposed to be a backing store for all
data in the app.

still some best practices/cleanup that need to be done, but currently
it’s a million times better than it was, and all old functionality
still works. we’ll see if I introduced any new problems later